- corected hang when there are electrical interferences on the arduino - corected bug on Shake forces - motor efficiency Min/Max instead of Left/Right

+++ Module1.vb
@@ -37,11 +37,11 @@
     End Function
 
 
-    Public Function ScaleValue(pValue As Single, pFromMin As Integer, pFromMax As Integer, pToMin As Integer, pToMax As Integer) As Integer
+    Public Function ScaleValue(pValue As Single, pFromMin As Single, pFromMax As Single, pToMin As Single, pToMax As Single) As Single
         If pValue = 0 Then Return 0
         Return Math.Min(pToMax, Math.Max(pToMin, (pValue - pFromMin) * pToMax / (pFromMax - pFromMin) + pToMin))
     End Function
-    Public Function ScaleValue(pValue As Single, pFromMin As Integer, pFromMax As Integer, pToMin As Integer, pToMax As Integer, pGama As Single) As Integer
+    Public Function ScaleValue(pValue As Single, pFromMin As Single, pFromMax As Single, pToMin As Single, pToMax As Single, pGama As Single) As Single
         If pValue = 0 Then Return 0
         Return (ScaleValue(pValue, pFromMin, pFromMax, pToMin, pToMax) / pToMax) ^ pGama * pToMax
     End Function
